### 2024-03-26 #6544 @kyleniemeyer
@openjournals/dev any idea why the paper build failed? The error seems to be
```
/opt/hostedtoolcache/Ruby/3.3.0/x64/lib/ruby/3.3.0/psych.rb:670:in `initialize': No such file or directory @ rb_sysopen - tmp/39b8d3331d64c1944c50eb44318968f3/paper/arxiv/paper.yml (Errno::ENOENT)
```
