### 2024-03-26 #6544 @xuanxu
I think the bot is trying to compile the paper using the paper.tex file as source, instead of using the paper.md, probably because the order they are listed when scanning the repository. Renaming the paper.tex file should make it work.
