### 2024-03-30 7d38ce9
remove empty author from the citation.cff
