+++ paper/arxiv/draft.tex
@@ -1,5 +1,7 @@
 \documentclass[a4paper,10pt]{amsart}
 
+\usepackage{amsaddr} % affiliation in the first page
+
 %%%%%%%%%%
 % PACKAGES
 %%%%%%%%%%
@@ -67,13 +69,12 @@
 \author{Thiago Araujo}
 
 \address{\noindent 
-Institute for Theoretical Physics, UNESP-Universidade Estadual Paulista
+Instituto de Física Teórica, UNESP-Universidade Estadual Paulista,
 R. Dr. Bento T. Ferraz 271, Bl. II, Sao Paulo 01140-070, SP, Brazil\\
-and
-Institute of Physics\\
-Universidade de S\~ao Paulo\\ 
-Rua do Matão Travessa 1371, 05508-090\\
-São Paulo, SP. Brazil}
+\&
+Instituto de Física, Universidade de S\~ao Paulo,
+Rua do Matão Travessa 1371, 05508-090 São Paulo, SP. Brazil
+}
 
 \email{\href{tr.araujo@unesp.br}{tr.araujo@unesp.br}}
 %\email{\href{tharaujo@if.usp.br}{tharaujo@if.usp.br}}
