+++ paper/paper.bib
@@ -1,23 +1,23 @@
 @article{Kelleher:2009,
-  author       = {Jerome Kelleher and
-                  Barry O'Sullivan},
-  title        = {Generating All Partitions: {A} Comparison Of Two Encodings},
-  journal      = {CoRR},
-  volume       = {abs/0909.2331},
-  year         = {2009},
-  url          = {http://arxiv.org/abs/0909.2331},
-  eprinttype    = {arXiv},
-  eprint       = {0909.2331},
+    author       = {Jerome Kelleher and Barry O'Sullivan},
+    title        = {Generating All Partitions: {A} Comparison Of Two Encodings},
+    journal      = {CoRR},
+    volume       = {abs/0909.2331},
+    year         = {2009},
+    url          = {http://arxiv.org/abs/0909.2331},
+    doi          = "10.48550/arXiv.0909.2331",
+    eprinttype   = {arXiv},
+    eprint       = {0909.2331},
 }
 
-
 @book{Macdonald:1998,
-  title={Symmetric Functions and Hall Polynomials},
-  author={Macdonald, I.G.},
-  isbn={9780198504504},
-  lccn={94027392},
-  year={1998},
-  publisher={Clarendon Press}
+    title={Symmetric Functions and Hall Polynomials},
+    author={Macdonald, I.G.},
+    isbn={9780198504504},
+    doi = "10.1093/oso/9780198534891.001.0001",
+    lccn={94027392},
+    year={1998},
+    publisher={Clarendon Press}
 }
 
 @book{Fulton:2004,
@@ -32,9 +32,10 @@
 
 @book{Babelon:2003,
 	title={Introduction to Classical Integrable Systems},
-  author={Babelon, O. and Bernard, D. and Talon, M. and Cambridge University
-  Press and Landshoff, P.V. and Nelson, D.R. and Sciama, D.W. and Weinberg,
-  S.},
+    author={Babelon, O. and Bernard, D. and Talon, M. and Cambridge University
+    Press and Landshoff, P.V. and Nelson, D.R. and Sciama, D.W. and Weinberg,
+    S.},
+    doi = "10.1017/CBO9780511535024",
 	isbn={9780521822671},
 	lccn={2002034955},
 	series={Cambridge Monographs on Mathematical Physics},
@@ -57,6 +58,7 @@
 @book{Marino:2005,
   title={Chern-Simons theory, matrix models, and topological strings},
   author={Marino, M.},
+  doi = "10.1093/acprof:oso/9780198568490.001.0001",
   isbn={9780198568490},
   lccn={2006272295},
   series={international series of monographs on physics},
@@ -69,6 +71,7 @@
     author = "Wheeler, Michael Alan",
     title = "{Free fermions in classical and quantum integrable models}",
     eprint = "1110.6703",
+    doi = "10.48550/arXiv.1110.6703",
     archivePrefix = "arXiv",
     primaryClass = "math-ph",
     school = "Melbourne U.",
