### 2024-04-18 #6544 @thraraujo
Hi @phibeck, thanks for pushing it forward.

- Some potential reviewers: szabo137, HugoStrand, Ethan-CS, eliotwrobson, AnnikaStein, Jpata
I don't know any of them, so there is not conflict of interests.
- DOIs have been added.
- I brought the paper down to 532 words.
