### 2024-04-19 #6544 @phibeck
:wave: @@szabo137, @HugoStrand & @eliotwrobson, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
