### 2024-04-19 #6544 @eliotwrobson
Happy to! I've wanted to teach myself a little bit about schur polynomials for a while and this would be a great excuse 😁
