+++ paper/arxiv/draft.tex
@@ -363,9 +363,9 @@ hall-littlewood: x1*x2*x3*(-Q**2*x1*x2*x3 - Q*x1*x2*x3 + x1**2*x2 + x1**2*x3 + x
 
 \section{Acknowledgements}
 
-I would like to thank Fapesp for finantial support, grant
-\textbf{2022/06599-0}.  I would also like to thank the Free and Open
-Source Software community.
+I would like to thank Fapesp for financial support, grant
+\textbf{2022/06599-0}.  I would also like to thank the Free and Open Source
+Software community.
 
 \printbibliography
      
