### 2024-05-06 #6724 @editorialbot
[REVIEW]: pySymmPol: Symmetric Polynomials
**Submitting author:** @thraraujo (<a href="http://orcid.org/0000-0001-5792-2530">Thiago Rocha Araujo</a>)
**Repository:** https://github.com/thraraujo/pysymmpol
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.2
