### 2024-05-08 thraraujo/pysymmpol#1 @AnnikaStein
[JOSS Review] Software / documentation
- [x] Syntax error in the last line of https://thraraujo.github.io/pysymmpol/getting-started.html#homogeneous-and-elementary-polynomials
- [x] https://thraraujo.github.io/pysymmpol/pysymmpol.html#module-pysymmpol Module contents is empty, do you need this heading then?

tbc, as the review progresses:

https://github.com/openjournals/joss-reviews/issues/6724
