### 2024-05-08 thraraujo/pysymmpol#1 @thraraujo
Hi @AnnikaStein, I've fixed a couple of typos in the documentation, and also corrected what you suggested here.
