### 2024-05-08 thraraujo/pysymmpol#2 @eliotwrobson
[JOSS Review] Type Annotations and CI
The code for this library looks pretty clean, but I think it would benefit from adding visible type annotations. Many functions have annotations already, and making these visible to client code would help usability of the package. For information on this, have a look here: https://blog.whtsky.me/tech/2021/dont-forget-py.typed-for-your-typed-python-package/

You should be able to do this just by making the new file, since poetry picks up this file by default.

Similarly, these annotations should be checked for correctness in the CI job, along with running the existing tests. For a short example of this in a GitHub workflow, see here (disclaimer: This comes from a project I worked on): https://github.com/eliotwrobson/CyNetDiff/blob/main/.github/workflows/tests.yml

This repo also seems like a nice example: https://github.com/LaurenceRawlings/python-poetry-cicd

From review https://github.com/openjournals/joss-reviews/issues/6724
