### 2024-05-08 thraraujo/pysymmpol#3 @eliotwrobson
[JOSS Review] Consistent internal representations and vectorized operations
Many of the internal data representations used in the library store coefficients using unannotated `tuple`s, see here https://github.com/thraraujo/pysymmpol/blob/main/pysymmpol/partitions/young.py#L26

You have numpy as a dependency, so I think it would be beneficial to switch the internal representation to use numpy vectors instead of tuples. Then, other internal operations can be implemented using fast vectorized operations in numpy, see: https://github.com/thraraujo/pysymmpol/blob/main/pysymmpol/polynomials/hall.py#L66

Similarly, library functions should accept numpy arrays as arguments to avoid unnecessary data conversions that can slow down these operations.

From review https://github.com/openjournals/joss-reviews/issues/6724
