+++ paper/paper.bib
@@ -77,3 +77,10 @@
     school = "Melbourne U.",
     year = "2010"
 }
+
+@manual{sagemath,
+  doi = "10.5281/zenodo.593563",
+  author       = {{The Sage Developers}},
+  title        = {{S}ageMath, the {S}age {M}athematics {S}oftware {S}ystem ({V}ersion 9.7)},
+  Year         = {2022},
+}
