### 2024-05-08 thraraujo/pysymmpol#1 @thraraujo
Hi @AnnikaStein, one last thing.

While I acknowledge the suggestion to test the software in other environments, I believe it might be overly cautious given the package's minimal dependencies. It's worth noting that the dependencies are not brand new releases.
