+++ .github/workflows/ci.yml
@@ -1,0 +1,26 @@
+name: Run Python Tests
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+    branches:
+      - master
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install Python 3
+        uses: actions/setup-python@v1
+        with:
+          python-version: 3.6
+
+      - name: Install dependencies
+        run: poetry install --with dev
+
+      - name: Run tests with pytest
+        run: poetry run pytest 
+      - name: Run mypy
+        run: poetry run mypy
