+++ pyproject.toml
@@ -14,6 +14,7 @@ numpy = "^1.26.4"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.1.1"
+mypy = "^1.10.0"
 
 [build-system]
 requires = ["poetry-core"]
