+++ .github/workflows/ci.yml
@@ -2,10 +2,10 @@ name: Run Python Tests
 on:
   push:
     branches:
-      - master
+      - dev
   pull_request:
     branches:
-      - master
+      - dev
 
 jobs:
   build:
