+++ .github/workflows/ci.yml
@@ -15,7 +15,7 @@ jobs:
       - name: Install Python 3
         uses: actions/setup-python@v1
         with:
-          python-version: 3.6
+          python-version: 3.10
 
       - name: Install dependencies
         run: poetry install --with dev
