+++ .github/workflows/ci.yml
@@ -1,21 +1,19 @@
 name: Run Python Tests
+
 on:
   push:
-    branches:
-      - dev
+    branches: ["*"]
   pull_request:
-    branches:
-      - dev
+    branches: ["*"]
 
 jobs:
   build:
     runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: Install Python 3
-        uses: actions/setup-python@v1
-        with:
-          python-version: 3.10
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
+        poetry-version: ["1.8.1"]
 
       - name: Install dependencies
         run: poetry install --with dev
