+++ .github/workflows/ci.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.9", "3.10", "3.11", "3.12"]
-        poetry-version: ["1.8.1"]
+        poetry-version: ["1.8.3"]
 
     steps:
       - uses: actions/checkout@v4
