+++ .github/workflows/ci.yml
@@ -26,6 +26,9 @@ jobs:
         with:
           poetry-version: ${{ matrix.poetry-version }}
 
+      - name: Install dependencies
+        run: poetry install --with dev
+
       - name: Run tests with pytest
         run: poetry run pytest 
 
