### 2024-05-08 thraraujo/pysymmpol#2 @thraraujo
Dear Eliot, thanks.

This is done, check the main branch. Github actions worked smoothly. See here: https://github.com/thraraujo/pysymmpol/actions/runs/9009994398
