### 2024-05-09 thraraujo/pysymmpol#2 @eliotwrobson
Output from new actions looks great! Will close.
