### 2024-05-09 thraraujo/pysymmpol#3 @thraraujo
I'm going to start working on this issue today. There are a couple of functions that I want to modify slightly, so I'll get started on that. It might take me 1 or 2 days to complete.

P.S. In the tests, there's a special case for the Schur functions tests where we calculate the Schur polynomials for 15 and 16 boxes twice each. This is why it's slow.
