### 2024-05-10 54e7a24
improveyoung diagram module
