### 2024-05-10 thraraujo/pysymmpol#1 @thraraujo
Dear @AnnikaStein, today I've checked the dependencies. It turns out that it works for python 3.11, sympy 1.11. I have not checked earlier versions. Numpy below 1.26.2 does not run. I have updated the README file and the documentation. It is in the main branch.
