### 2024-05-10 2515a1a
fix young diagram module
