### 2024-05-10 a7c4ec4
improve young and conjugacy modules
