### 2024-05-11 thraraujo/pysymmpol#3 @thraraujo
Dear @eliotwrobson,

Thanks for your insightful suggestions; I have worked on them.

1. I've decided to maintain tuples as the default structure for user interaction. While internally utilizing numpy methods, the package's interface remains primarily pure Python, with potential integration of sympy. I appreciate the immutability and the clarity of input and output formats that tuples offer. However, I've heeded your advice to some extent by now allowing the use of numpy arrays.

2. Your point about utilizing numpy arrays for performance enhancement is well taken. I've begun integrating numpy arrays in certain sections of the code, such as the transpose method and within the states module. The latter, being pivotal to the package's functionality in physics, now incorporates numpy arrays in strategic areas. Additionally, I'm considering reimplementing this module in a faster language for long-term efficiency goals, aiming to handle all combinatorics in C while utilizing the package for symbolic polynomial manipulations.

3. I've optimized some methods to minimize nested 'for loops'. For instance, the method for calculating diagonal elements now employs arrays and requires only one 'for loop.' Similarly, enhancements have been made to the contains and interlaces methods within the Young module, significantly reducing computational overhead by terminating at the first encountered false condition rather than iterating through all possibilities.

4. Regarding the Hall-Littlewood polynomials example you mentioned, I adopt dictionaries instead of tuples to address efficiency concerns. Although it slightly streamlined the code and improved readability, it poses algorithmic challenges, particularly in efficiently iterating over coordinates and permutations. While I currently favor this approach, I remain open to exploring alternative strategies, especially as I consider implementing other polynomial types like MacDonald and Jack **in the future**.

5. I've updated the readme to include valuable references on relevant topics, including the books you recommended.
