### 2024-05-11 thraraujo/pysymmpol#3 @eliotwrobson
@thraraujo your changes look awesome, thanks for addressing them so quickly! For context, this issue was mainly meant to draw your attention to potential easy performance gains (especially if your package gets used on larger inputs), rather than a call to do a larger refactoring right away, and I think the changes you have incorporated have more than capitalized on that so far. This was never meant to be a review blocker either, so I will go ahead and close. One short comment on one of your items:

> Additionally, I'm considering reimplementing this module in a faster language for long-term efficiency goals, aiming to handle all combinatorics in C while utilizing the package for symbolic polynomial manipulations.

This is actually the reason I mentioned using numpy arrays in more places in the first place. Instead of reimplementing in C, if performance becomes an issue, you could consider converting substantial parts of your package to Cython and adding a compilation step to the package building: https://cython.readthedocs.io/en/latest/src/tutorial/numpy.html

Cython integrates very cleanly with code that heavily uses Numpy, so it makes sense to have an eye towards using it where possible now if that conversion is something you'd be interested in later. As they say, premature optimization is the root of all evil, so I'd suggest not taking the plunge on this quite yet if you don't find yourself needing it.
