### 2024-05-11 thraraujo/pysymmpol#4 @eliotwrobson
[JOSS Review] Minor Nitpicks
Just some small miscellaneous things I noticed:

- [ ] On this line, the check can be changed to `isinstance(self.partition, (tuple, np.ndarray))`: https://github.com/thraraujo/pysymmpol/blob/main/pysymmpol/partitions/young.py#L33

See https://stackoverflow.com/a/33311330
- [ ] Similarly, the `all` function can be called on a generator expression instead of a list, and using the slice here makes a copy: https://github.com/thraraujo/pysymmpol/blob/main/pysymmpol/partitions/young.py#L39

It would be better to have `A = all(x >= y for x,y in pairwise(par))`, avoiding the issues mentioned above.

- [ ] In this same file (and other places), you use the raw `tuple` type annotation, when you really only want tuples that contain numbers. This can be done with the annotation `tuple[numbers.Number, ...]`, indicating an indeterminate length tuple containing only numbers: https://stackoverflow.com/a/4187266

Similarly specific annotations can be done with a lot of the `dict` annotations in the package as well.

- [ ] In the `count_diagonal` function, you create a whole matrix and then compute the trace: https://github.com/thraraujo/pysymmpol/blob/main/pysymmpol/partitions/young.py#L88

However, all you are doing in the inner loop is constructing a row of `partition[n]` ones and padding with zeros, which is wasteful for large diagrams. Instead, you can just iterate through each row of the diagram and add 1 to your final result if `partition[n] >= n`, indicating that there would be a 1 on the diagonal. In fact, I think this is done by the one-liner `sum(1 if part >= n else 0 for n, part in enumerate(partition))`, as this adds 1 for each row in the diagram that crosses the diagonal.

- [ ] I like your statement of need from the paper you submitted, but this should be made part of the documentation + README.md as well.

Part of the ongoing JOSS review: https://github.com/openjournals/joss-reviews/issues/6724
