### 2024-05-11 thraraujo/pysymmpol#4 @thraraujo
Hi @eliotwrobson,

Thank you once again. Regarding the other issue, I've made significant updates to the count diagonal and Frobenius coordinates methods. Now, I've completely revamped the approach. Initially, I compute the Frobenius coordinates through a loop, utilizing both physicist and mathematician definitions. Once these expressions are derived, I proceed to calculate the diagonal terms based on the length of these coordinates. Additionally, I've implemented some tests to ensure accuracy, and I verify the number of boxes in the diagonal using hook lengths to validate the results.

I'll revisit the remaining points on Monday. Thanks again
