### 2024-05-11 thraraujo/pysymmpol#4 @eliotwrobson
Awesome! I think the new code computing this recomputes the transpose repeatedly in the inner loop, so you should store this in a variable if you can, but otherwise looks much cleaner that before 👍👌
