### 2024-05-11 819de7f
improve frobenius coord method
