+++ .github/workflows/ci.yml
@@ -2,9 +2,9 @@ name: Run Python Tests
 
 on:
   push:
-    branches: ["*"]
+    branches: ["dev"]
   pull_request:
-    branches: ["*"]
+    branches: ["dev"]
 
 jobs:
   build:
