### 2024-05-11 thraraujo/pysymmpol#4 @thraraujo
well noticed! fixed.
