### 2024-05-13 2946f6a
implement a better method to calculate the transposition
