### 2024-05-13 c3b5029
fix hall littlewood equation
