### 2024-05-13 f57c31b
change less than to greater than in hall littlewood
