### 2024-05-13 d910a7d
change less than to greater than in hall littlewood - one more try
