### 2024-05-13 thraraujo/pysymmpol#4 @thraraujo
Hi @eliotwrobson,

I wanted to share some (in my opinion) positive updates:

1. Upon reviewing volume 4 of TAOCP, I came across an interesting relationship between the conjugate of a Young diagram and its conjugacy representation. Leveraging this connection, I've revamped the method for calculating the transpose, resulting in a significant improvement.

2. Additionally, I've addressed the other points you raised. Thank you for those suggestions; I found them quite beneficial, especially the 'pairwise' tip.

3. I might have overlooked something, but I do agree with the necessity of validating the type. I've implemented it differently, though.

4. The review has been included in the readme. However, I haven't yet added it to the documentation. I plan to do so in the future; I just need to figure out how to incorporate LaTeX support into Sphinx. It's likely straightforward, but I prefer not to delve into that at the moment.
