### 2024-05-13 thraraujo/pysymmpol#4 @eliotwrobson
@thraraujo that all sounds excellent! None of these items were hard review blockers (besides the statement of need), just things I noticed that could possibly clean up parts of the package. If you're satisfied with changes made based on my suggestions, feel free to close this issue 😃
