### 2024-05-13 thraraujo/pysymmpol#2 @eliotwrobson
@thraraujo I think the `py.typed` file needs to be in the `pysymmpol` folder, as this is the package directory (while the root of the repository is not).
