### 2024-05-13 thraraujo/pysymmpol#4 @thraraujo
@eliotwrobson, thanks... In fact, I was just checking the statement of need. I didn't like the way sphinx adds math support, so I created a link to the repo. Case closed!
