### 2024-05-14 thraraujo/pysymmpol#6 @eliotwrobson
[JOSS Review] Documentation
The documentation includes the functions and classes in the `tests` folder, even though these are not part of the package. These should be removed: https://thraraujo.github.io/pysymmpol/tests.html

Additionally, the test uses the unittest class inheritance structure, instead of the just the cleaner naming convention traditionally used by pytest (no need for the wrapper classes). However, feel free to ignore this since it requires changing the way assertions need to be done, but still something to keep in mind.

Part of https://github.com/openjournals/joss-reviews/issues/6724
