### 2024-05-14 thraraujo/pysymmpol#1 @AnnikaStein
Very nice, thank you for working on these! Feel free to close this issue.
