### 2024-05-14 thraraujo/pysymmpol#6 @thraraujo
Hi @eliotwrobson

I removed the test from the documentation.

Regarding the second point; some of these tests were implemented before incorporating pytest into my workflow. For consistency's sake, I'd rather not retroactively modify these tests. In future projects I can use pytest standards from the outset.
