### 2024-05-14 thraraujo/pysymmpol#6 @eliotwrobson
@thraraujo looks good! As I said, the test format changes were not blocking, so I will go ahead and close.
