+++ paper/paper.bib
@@ -84,3 +84,13 @@
   title        = {{S}ageMath, the {S}age {M}athematics {S}oftware {S}ystem ({V}ersion 9.7)},
   Year         = {2022},
 }
+
+@inproceedings{Okounkov:2006,
+   title={The uses of random partitions},
+   url={http://dx.doi.org/10.1142/9789812704016_0037},
+   DOI={10.1142/9789812704016_0037},
+   booktitle={XIVth International Congress on Mathematical Physics},
+   publisher={WORLD SCIENTIFIC},
+   author={OKOUNKOV, ANDREI},
+   year={2006},
+   month=mar }
