### 2024-05-14 #6724 @thraraujo
Hi @eliotwrobson,

I'll address the last issue here as it's pertinent to concerns raised by @AnnikaStein. It's worth noting that both @phibeck and @AnnikaStein are fellow physicists.

**Regarding your comment:**
> _One of the review checklist items is a discussion of the "state of the field" about other software packages. Some discussion of this would be very useful, even if it is just stating what operations are supported by this package that do not exist elsewhere (i.e. not in sympy)._

- I've incorporated another section on the state of the field, delineating the main disparities between our package and SageMath. Additionally, I've included insights on the target audience.

**Regarding your comment:**
>  _The section on the core features of the package would be much stronger if there were an example of the analysis that could be done with the package in relation to the discussion of CFTs. As someone with a pure combinatorics background, it's a little hard to tell what types of analysis can be done with the package in the context of theoretical physics._

- Describing the types of analyses possible isn't straightforward, given the myriad combinatorial problems in physics. However, I've appended references to Okounkov's and Marino's papers, which align with my interests. In any case, you might check examples ranging from condensed matter physics [Zinn-Justin](https://arxiv.org/pdf/2404.13221) to string theory [Nekrasov-Okounkov](https://arxiv.org/pdf/hep-th/0306238).

- One relevant example within your interests is in representation theory. Quantum mechanics operators (representing measurable objects) are unitary (Lie groups SU(n)), and through the [Schur-Weyl duality](https://en.wikipedia.org/wiki/Schur%E2%80%93Weyl_duality), their representations are labeled by Young diagrams, related to symmetric group representations. Schur polynomials serve as [characters](https://en.wikipedia.org/wiki/Schur_polynomial#Relation_to_representation_theory) in these representations.

- While I refrain from self-citation, I've conducted research leveraging this software to examine objects called [tau functions](https://en.wikipedia.org/wiki/Tau_function_(integrable_systems)#Fermionic_VEV_(vacuum_expectation_value)_representations), as elaborated in this [paper](https://inspirehep.net/literature/2782843). These objects were expanded in terms of Schur and Hall-Littlewood polynomials.

- Lastly, I aim to keep the paper concise, adhering to <700 words, as per @phibeck's [comments](https://github.com/openjournals/joss-reviews/issues/6544#issuecomment-2064102284).
