### 2024-05-16 #6724 @thraraujo
Hi @eliotwrobson and @AnnikaStein, thanks for all the comments.

Hi @phibeck, the only missing point appears to be the ***Summary*** in @AnnikaStein's checklist. @eliotwrobson's checklist agrees that the summary already provides a high-level description understandable even to non-specialists. In my opinion, key terms like ***Manipulation***, ***Polynomials***, and ***Symmetric*** are very precise, so I don't want to give an additional high-level description of the types of manipulations or symmetric polynomials the package deals with. Since this is a minor point, I'll leave the decision to you.
