### 2024-05-16 #6724 @AnnikaStein
Sorry @thraraujo for not being clear on what I meant with this point - I was a bit hesitant to check it as the summary contains abbreviations, "AdS/CFT" which, in my opinion, a "diverse, non-specialist audience" would likely not understand without further reading. For a someone with a physics background, one may always infer that field theories are meant when you see the letters FT next to each other :), but if one looks at it from a different perspective, it's a non-trivial set of abbreviations.
