### 2024-05-16 a6bbdc7
explain what ads cft mean
