+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pysymmpol"
-version = "0.1.1"
+version = "0.1.2"
 description = "A python package for manipulation of symmetric polynomials."
 authors = ["Thiago Araujo"]
 license = "GPL-3.0"
