### 2024-05-24 #6724 @thraraujo
Hi @phibeck , thanks

> Thank you very much for an incredibly speedy and thorough review, @AnnikaStein and @eliotwrobson! 🚀 And thank you @thraraujo for addressing all comments timely and adequately. 🙌 This is by far the fastest review I have seen so far, thank you all! 🥇

Kudos to @AnnikaStein and @eliotwrobson! They were indeed very fast, and I sincerely believe the package is much better now.

> - [X]  Make a tagged release of your software, and list the version tag of the archived version here.

I have already done this, version **v0.1.2**

>  - [X]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

Archived here: https://zenodo.org/records/11214737

> - [X]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

Done.

> - [X]  Please list the DOI of the archived version here.

DOI: [10.5281/zenodo.11214737](https://doi.org/10.5281/zenodo.11214737)
