### 2024-05-28 #6724 @phibeck
@thraraujo thanks for the reminder. I'm away for a workshop since last week so response is a bit slow. Here are a few more comments/suggestions for the manuscript. Please have a look when you have a moment.

- line 20f: I assume the four references are associated with the four fields mentioned here, but the association is a bit ambiguous. Could you place the reference directly with the mentioned field?
- line 36: CFT is already defined in line 14, so I think it's not necessary here
- line 36f: when citing multiple references at once, citing them as `Blah blah [@doe99; @smith2000; @smith2004]` should render them as one block (see https://pandoc.org/MANUAL.html#extension-citations)
- line 37: "for examples of physical problems..."
- line 78: Hall capitalized
- line 80: Simons capitalized
