### 2024-05-28 344c39f
correct bib entries
