+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @book{Macdonald:1998,
-    title={Symmetric Functions and Hall Polynomials},
+    title={Symmetric Functions and {H}all Polynomials},
     author={Macdonald, I.G.},
     isbn={9780198504504},
     doi = "10.1093/oso/9780198534891.001.0001",
@@ -56,7 +56,7 @@
 }
 
 @book{Marino:2005,
-  title={Chern-Simons theory, matrix models, and topological strings},
+  title={{C}hern-{S}imons theory, matrix models, and topological strings},
   author={Marino, M.},
   doi = "10.1093/acprof:oso/9780198568490.001.0001",
   isbn={9780198568490},
@@ -90,7 +90,7 @@
    url={http://dx.doi.org/10.1142/9789812704016_0037},
    DOI={10.1142/9789812704016_0037},
    booktitle={XIVth International Congress on Mathematical Physics},
-   publisher={WORLD SCIENTIFIC},
-   author={OKOUNKOV, ANDREI},
+   publisher={World Scientific},
+   author={Okounkov, Andrei},
    year={2006},
    month=mar }
